[Important] debug issues
In previous exprements, using cp [file] to /dev/scullc and cat /dev/scullc
to see if driver is successfully work. But cp changed the /dev/scullc
file to be a normal file, mislead my judging.
This issue is found when debug the PDEBUG macro.

Another issue is the CFLAGS variable is not used when compiling codes.
Instead using EXTRA_CFLAGS makes the compiling options effective.